LMS Autocomplete backspace changes and Timesheet leave days count to sum

+++ NLTD.EmployeePortal.LMS.Common/DisplayModel/TimeSheetConsolidate.cs
@@ -12,12 +12,12 @@ namespace NLTD.EmployeePortal.LMS.Common.DisplayModel
         public TimeSpan hours { get; set; }
         
         public int permissionCount { get; set; }
-        public int leaveCount { get; set; }
+        public decimal leaveCount { get; set; }
         public int lateCount { get; set; }
         public int earlyCount { get; set; }
         
        public DateTime fromDatetime { get; set; }
         public DateTime toDatetime { get; set; }
-        public int WFHCount { get; set; }
+        public decimal WFHCount { get; set; }
     }
 }
